+++ CHANGELOG.md
@@ -58,11 +58,15 @@
 - BPM workflow dashboard: Updated chart titles and fixed issue with per day status chart. 
 - Configure BoM form: show sub bom list.
 - Configurator: improve UI for indicators.
+- Split formula list into formulas for product and formulas for sale order line in configurator creator.
+- Add a column name to configurators, equal to the creator's name.
 
 ## Bug Fixes
 - Fix on prod process report.
 - Fix json field creator's issue of model change not reflected on fields. 
 - Configurator export: add contextField to metaJsonField.
+- Configurator creator: fill default attributes for both product and sale order line.
+- Computation of value of a given stock location is now the same in the form and in the printing.
 
 
 ## [5.0.4] - 2018-07-26